chore: map rendering optimization and cleanup (#2443)
* chore: initial cleanup of lookups/map rendering

Unfortunately this has yet to result in a noticeable performance increase despite what should be a reduction in lookup accesses and mutex locking...

* chore: slight reduction of wasteful calls

* chore: more code cleanup, map rendering optimization

* chore: lookup/access cleanup

* fix: better error messages, member names

* chore: more optimizations and clean up

* fix: remove table/label invalidation loops

* fix: prevent KeyNotFoundException and any potential OOBs

* chore: labled constants instead of unlabeled magic numbers

+++ Intersect.Client.Framework/Maps/IActionMessage.cs
@@ -7,7 +7,7 @@ public interface IActionMessage
     string Msg { get; set; }
     long TransmissionTimer { get; set; }
     int X { get; set; }
-    long XOffset { get; set; }
+    int XOffset { get; set; }
     int Y { get; set; }
 
     void TryRemove();
